Split bdev into modules: lxcloop
The functions

	- loop_clonepaths();
	- loop_create();
	- loop_destroy();
	- loop_detect();
	- loop_mount();
	- loop_umount();

move from bdev.c to lxcloop.{c,h}. All functions previously declared static
become extern.

The functions

	- do_loop_create();
	- find_free_loopdev_no_control();
	- find_free_loopdev();

move from bdev.c to lxcloop.c. They remain static.

Adapt Makefile.am to include lxcloop.{c,h}.

The structs

       - struct bdev; /* defined in bdev.h */
       - struct bdev_specs; /* defined in lxccontainer.h */
       - struct lxc_conf; /* defined conf.h */

are forward declared/put as incomplete types into lxcloop.h as the functions
associated with loop need access to it.
